+++ config/config_manager.go
@@ -109,6 +109,7 @@ func (cm *ConfigManager) setConfigDefaults(v *viper.Viper) {
 	v.SetDefault("auth.issuer", "gosso")
 	v.SetDefault("auth.authorization_code_expiry", "5m")
 	v.SetDefault("auth.device_code_expiry", "10m")
+	v.SetDefault("auth.device_code_interval", "5s")
 
 	// Log configuration
 	v.SetDefault("log.level", 0)